vendor_management/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4,<4.0
drf-yasg>=1.20.0,<2.0
djangorestframework-simplejwt>=4.8.0,<5.0